requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic~=2.10.3
pydantic-settings==2.6.1
face-recognition==1.3.0
pillow==11.0.0
numpy==2.1.3
python-multipart==0.0.6
pytest~=8.3.4
starlette~=0.41.3